API/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql.connector
mysql-connector-python
aiohttp